clean up unused component and code
