--- addon/templates/components/frost-related-details.hbs
@@ -1,15 +1,0 @@
-{{#if (eq myTargetRouteName route)}}
-    <div class='circle'>
-      {{#if svgPath}}
-        {{frost-icon icon=svgPath}}
-      {{/if}}
-    </div>
-    {{yield}}
-{{else}}
-    <div class='circle'>
-      {{frost-icon icon='frost/close'}}
-    </div>
-    {{yield}}
-{{/if}}
-
-
