--- app/components/frost-related-details.js
@@ -1,1 +1,0 @@
-export {default} from 'ember-frost-object-details/components/frost-related-details'
